+++ R/rprojroot-package.R
@@ -1,0 +1,15 @@
+#' @details
+#' See the "Value" section in [root_criterion()] for documentation
+#' of root criterion objects, and  [criteria()] for useful predefined
+#' root criteria.
+#'
+#' @examples
+#' criteria
+#' \dontrun{
+#' is_r_package$find_file("NAMESPACE")
+#' root_fun <- is_r_package$make_fix_file()
+#' root_fun("NAMESPACE")
+#' }
+#' @import backports
+#' @api
+"_PACKAGE"